Fix a float-to-int deprecation in antispambot() surfaced by the feed export
EN:
antispambot() seeded the RNG with `srand( (float) microtime() * 1000000 )`.
The argument is a float, and the implicit float-to-int conversion is a
deprecation in PHP 8.1+ ("loses precision"). It fired whenever the feeds
(b2rss2.php / b2rdf.php) obfuscate an author email, and the static export
baked the deprecation text into the exported feed files. Cast the seed to int.

JA:
antispambot() は `srand( (float) microtime() * 1000000 )` で乱数を初期化して
いた。引数は float であり、暗黙の float→int 変換は PHP 8.1 以降で非推奨
(「精度が失われる」)。フィード(b2rss2.php / b2rdf.php)が著者メールを
難読化するたびに発火し、静的書き出しがその非推奨テキストをフィードファイル
に焼き込んでいた。シードを int にキャストする。

Refs #57

+++ src/b2-include/b2functions.php
@@ -331,7 +331,11 @@ function convert_smilies( $content ) {
 
 function antispambot( $emailaddy, $mailto = 0 ) {
 	$emailNOSPAMaddy = '';
-	srand( (float) microtime() * 1000000 );
+	// EN: cast the seed to int; an implicit float-to-int conversion is a
+	//     deprecation in PHP 8.1+ ("loses precision").
+	// JA: シードを int にキャストする。暗黙の float→int 変換は PHP 8.1 以降
+	//     非推奨(「精度が失われる」)。
+	srand( (int) ( (float) microtime() * 1000000 ) );
 	for ( $i = 0; $i < strlen( $emailaddy ); $i = $i + 1 ) {
 		$j = floor( rand( 0, 1 + $mailto ) );
 		if ( 0 == $j ) {
